backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.4.2
factory-boy==3.3.0
Faker==20.1.0
freezegun==1.3.1
kombu==5.3.4
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg==3.1.14
psycopg-binary==3.1.14
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.42.0
